# api/openapi.yaml
openapi: 3.0.0
info:
  title: Calculator API
  description: API for basic math operations with JWT authentication
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /register:
    post:
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
            example:
              email: user@example.com
              password: password123
              name: John Doe
      responses:
        '201':
          description: User created
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-swagger-router-controller: Default

  /login:
    post:
      summary: Login user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
            example:
              email: user@example.com
              password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-swagger-router-controller: Default

  /calculate:
    post:
      summary: Perform a mathematical operation
      operationId: calculate
      parameters:
        - name: operation
          in: header
          description: The operation to perform
          required: true
          schema:
            type: string
            enum: [add, subtract, multiply, divide]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateBody'
            example:
              number1: 10
              number2: 5
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResult'
        '400':
          description: Invalid input or missing operation header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - JWT: []
      x-swagger-router-controller: Default

components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication

  schemas:
    UserRegister:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string

    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    CalculateBody:
      type: object
      required: [number1, number2]
      properties:
        number1:
          type: number
        number2:
          type: number

    CalculationResult:
      type: object
      properties:
        result:
          type: number
        userId:
          type: integer
        operation:
          type: string
        operands:
          type: array
          items:
            type: number

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      example:
        message: "Invalid input: number1 must be a number"